requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.1.4
django-debug-toolbar==4.4.6
django-heroku==0.3.1
django-import-export==4.3.3
django-model-utils==5.0.0
django-simple-history==3.7.0
django-tinymce==4.1.0
django-unfold==0.43.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==24.2
pathlib==1.0.1
pillow==11.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycryptodome==3.21.0
pyserial==3.5
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.2
tablib==3.7.0
tenacity==5.1.5
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.8.2
